requirements.txt
aiogram
aiohttp
supabase
python-dotenv
shortuuid
